+++ internal/accrual/storage/migrations/000002_create_order_items_table.up.sql
@@ -1,5 +1,5 @@
 create table if not exists order_items(
-    order_id int not null,
+    order_id bigint not null,
     description text not null,
     price decimal not null,
     constraint fk_order_id foreign key(order_id) references orders(id)
